add a 'object.failed_deposit' event
when a deposit fails, we can publish an event; this gets that in place, in
anticipation of publishing them from the `WorksControllerBehavior`, and possibly
other places (batch)?

this is extracted from some exploratory work on using our transactions interface
for Valkyrie work deposit.

some consideration about how to document Hyrax's pub/sub contracts is still
owed.