Removing Capybara Maleficent
Jeremy here, I introduced Maleficent at Notre Dame as we observed odd
behavior in an older version of Capybara.  I know for certain that
Maleficent is not keeping up with Capybara releases.

At the time of its creation, there were assertions that Capybara was
waiting for DOM resolution.  However, in our observations this was not
always true.

However, this is a place where we could very easily be experiencing
increased fragility (and a slow down in specs).

I'd like to see if things work without this past hack.

NOTE: Today, I have been walking back gems from recently released
versions.  It's slow going, and my observation is that the feature specs
are the fragile ones.  The inner workings of Capybara may have changed
and thus either: 1) removing the need of Maleficent or 2) breaking
Maleficent.  So let's remove Maleficent and see how things fare.