Force linkeddata 3.x for compatibility with rdf 3.x
I ran into this issue when testing where ebnf 2.0 was released on Friday and was getting pulled in.  The only version of linkeddata which resolves then is 1.1.1 which has compatibility issues with rdf 3.x.  If ebnf is backed down to 1.x then I believe the correct version of linkeddata will be used but specifying the version of linkeddata better describes the true dependency requirement.