Removing magic numbers
Prior to this commit, there was a presumption about the order in which
things ran AND the state of the auto-incrementer for the database.

This works, however, when you move to a different
database (e.g. Postgresql), this behavior won't work (unless you do some
really expensive work between each test).

Instead of relying on magic numbers, provide the number based on the
object.