Reworking method call to improve clarity
Prior to this commit, it wasn't immediately evident that the
`curation_concern_from_search_result` was meant to run for a single
item (e.g., the given `params[:id]`).  It becomes evident when you look
at the behavior of `search_result_document`, which takes the first item
OR fallsback to a `SolrDocument.find(params[:id])`

In addition, this removes a mutation on the params which COULD have some
odd behavior.  Again, given that we're working with a singular object, I
don't think :page is a likely parameter.

I also chose to write a comment to help capture what I've discovered in
some additional work.  It's an odd thing and I want to make sure other
see it.