rearrange route stubs for WorksControllerBehavior mixin tests
the mixin tests on WorksControllerBehavior are meant to be unit tests, so we
want to test against a minimal set of required collaborators (including
routes/url helpers). in this case, we need some routes defined on the "main
app". to ensure they're there, we stub the `#main_app` helper.

meanwhile, use a simple stub for registering the curation concern, avoiding the
need for complex `before/after(:context)` test setup. this should reduce
order-dependent test failures.