Adding .inspect to anonymous WorkSearchBuilder
Prior to this commit, when working with the anonymous search builder, it
was difficult to see exactly how the search builder ended up configured.

This change now provides insight both to the original work type and the
legacy work type.  This should help in debugging.

In particular, I've found it useful in debugging the mixin module:
`app/controllers/concerns/hyrax/works_controller_behavior.rb`

Tested-by: Jeremy Friesen <jeremy.n.friesen@gmail.com>