Isolating variables for WorkSearchBuilder
Prior to this commit, the module variables set in the anonymous class
were shared across classes built via `Wings.WorkSearchBuilder`.

I believe this may be a key contributor to erratic spec failure.

I observed this when I was running specs and saw that the resulting
search builder would query for work_types of "Wings(Monograph)" but the
indexed object had a model of "Hyrax::Test::SimpleWorkLegacy".  The
"Wings(Monograph)" and "Hyrax::Test::SimpleWorkLegacy" are not expected
to map to each other in the `Wings::ModelRegistry.lookup`.

If I ran the specs in isolation, they would always pass.  However, if I
ran the suite, I would often see failures for the specs that passed in
isolation.

`rspec
./spec/controllers/concerns/hyrax/works_controller_behavior_spec.rb`

By switching to [.class_attribute][1], the variables are inheritable and
changes in subclasses don't impact changes upstream.

[1]:https://api.rubyonrails.org/classes/Class.html#method-i-class_attribute

Here's hoping!