prepopulate forms before rendering
reform supports a `#prepopulate!` method for configurable defaults within the
form object. our base form objects use this to provide a `#version` (optimistic
lock) token, as well as permissions and embargo/lease data (which don't travel
with the object). ensuring we call it before displaying the form ensures these
fields are populated as expected. it also makes it possible for adopters to
extend with their own preopulator logic, supporting a wide range of
default-value behavior for forms.