Adjusting UsersController#search for agnosticism
Prior to this commit, I believe the syntax was less generalize.  In
particular, the "field like lower(value)" was suspect. Namely, we were
already downcasing the query, but we weren't telling the field to be
treated as lower case.

I believe in SQLite and Mysql, LIKE clauses are already assumed to be
case agnostic.  That is, I think, not the case in Postgresql.  In other
words, Postgresql needs to be told whether or not to treat the field as
lower (or upper) case.

With this change, I'm also introducing sanitization of possible "like"
characters as part of the query.

And, I also note that this method has feature envy; That is to say the
method sends lots of messages to `::User`, strongly implying that this
method belongs in `::User`