Adjusting specs to be less brittle
Prior to this commit, there is an assumption that the database
auto-incrementer was always reset to 1 between tests.  This is true with
Sqlite, but not true with other databases.  So, instead of using a magic
number, I'm opting to use the ID of the object that was created for the
spec.

In the case of `eq(agents)` and `contain_exactly(*agents)`, this is a
similar-ish concern.  When using `eq` the order of elements matter.  And
the order of those elements can vary from DB adapter.  Instead, rely on
the more flexibily `contain_exactly`.