Fixing parameter order for `have_content`
The following are [specs copied from Capybara 3.33.0][1]

```ruby
it 'can check for all text' do
  expect(session).to have_text(:all, 'Some of this text is hidden!')
end

it 'can check for visible text' do
  expect(session).to have_text(:visible, 'Some of this text is')
  expect(session).not_to have_text(:visible, 'Some of this text is hidden!')
end
```

This change complies with the parameter order identified in Capybara.
Note "have_content" is an alias for "have_text" ([see code][2])

Background:

I noticed that there was an [odd error in our CircleCI build][3]; This
was both unchanged and not something we specifically looked at.  Based
on cursory exploration, this appears to have been the method format
since at least 2.3.0.  Why it's failing now?  No clue.  But hopefully
this fixes that particular failure.

[1]:https://github.com/teamcapybara/capybara/blob/3.33.0/spec/rspec/shared_spec_matchers.rb#L459-L466
[2]:https://github.com/teamcapybara/capybara/blob/3.33.0/lib/capybara/rspec/matchers.rb#L127-L133
[3]:https://app.circleci.com/pipelines/github/samvera/hyrax/3692/workflows/2998f493-989c-45d0-806a-5d397c624140/jobs/24460/steps