Ensuring that we have a default admin set
In introducing clean_repo logic to all feature tests, this feature spec
started regularly failing.  In the past, it would fail somewhat
arbitrarily.

With this change, we're ensuring that we have a default admin set before
running the spec.  Prior to the change, we were instatiating the default
admin set but not persisting it.  This leads me to believe that the spec
succeeded when a prior run spec would persist and admin set AND no
clean_repo calls were made between that prior spec and this spec.

As writing, just building the admin_set does not make it available to
feature tests (as the feature will be querying for persisted objects).