Refactoring specs to ignore array order
Prior to this commit, it was possible that the Rspec `.with` stubs would
receive collection type gids in an order different from what was
expected.  I believe this is a Postgresql only issue, as I have memories
that Postgresql does not apply an auto-sort by ID.  That may be a faulty
memory, however, the result is a fragile spec in which we have an
explicit order in the `.with` but a stochastic order in the actual array
that is built from a production code query.

In addition, I have take the opportunity to rename the variables and
methods to be more descriptive and precise.  I would expect variables
and method names that say "single_membership_types" to be a model (e.g.,
`Hyrax::CollectionType`) but they were instead an array of GIDs.

This change also draws the code closer to dependency injection, in that
we remove a chained call in favor of a method call.

Below is the failure that [I observed in CircleCI][1]:

```console
Failure/Error: ::Collection.where(id: collection_ids,
collection_type_gid_ssim: single_membership_types)
  <Collection (class)> received :where with unexpected arguments
  expected: ({:collection_type_gid_ssim=>["gid://internal/hyrax-collectiontype/52", "gid://internal/hyrax-collectiontype/53"], :id=>["object_id_8", "object_id_9"]})
  got: ({:collection_type_gid_ssim=>["gid://internal/hyrax-collectiontype/53", "gid://internal/hyrax-collectiontype/52"], :id=>["object_id_8", "object_id_9"]})
  Diff:
  @@ -1,5 +1,5 @@
  [{:collection_type_gid_ssim=>
  -   ["gid://internal/hyrax-collectiontype/52",
  -    "gid://internal/hyrax-collectiontype/53"],
  +   ["gid://internal/hyrax-collectiontype/53",
  +    "gid://internal/hyrax-collectiontype/52"],
     :id=>["object_id_8", "object_id_9"]}]
```

[1]:https://app.circleci.com/pipelines/github/samvera/hyrax/3714/workflows/e08b0c17-eb26-4e10-99ff-c02ea7ca6f78/jobs/24737/steps