Refactoring proxy feature spec
Prior to this commit, there were two expensive feature specs.  The first
spec (e.g. "creates a proxy") is entirely a subset of a second feature
spec that created a proxy, deleted the proxy, and verified the deletion.

In addition, the second feature spec signed the same user in twice.  I
believe the reason being that we wanted to ensure that we were at the
same URL for the second set of tests.

I have added a temporary variable to capture that URL and avoid calling
a second time the computationally expensive sign-out/sign-in helper.

My hope is that this will address the following [CircleCI error][1]:

```console
1) proxy add proxy in profile delete a proxy, and verify it was deleted
   Failure/Error: fill_in 'Email', with: user.email

Capybara::ElementNotFound:
   Unable to find field "Email" that is not disabled

   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/node/finders.rb:302:in `block in synced_resolve'
   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/node/base.rb:83:in `synchronize'
   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/node/finders.rb:291:in `synced_resolve'
   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/node/finders.rb:52:in `find'
   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/node/actions.rb:91:in `fill_in'
   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/session.rb:759:in `block (2 levels) in <class:Session>'
   ./vendor/bundle/gems/capybara-3.33.0/lib/capybara/dsl.rb:58:in `block (2 levels) in <module:DSL>'
   ./spec/support/features/session_helpers.rb:9:in `sign_in'
   ./spec/features/proxy_spec.rb:47:in `block (3 levels) in <top (required)>'
 ```

[1]:https://app.circleci.com/pipelines/github/samvera/hyrax/3714/workflows/3ad37bcd-e08b-47d8-b6b8-0982ef89c591/jobs/24731/steps