publish fixity events directly; drop reliance on deprecated callback
we're deprecating all the Hyrax "callbacks" in favor of the new pub/sub
interface. the behavior of the old callback has already moved to a listener.
instead of relying on the callback to publish the event, publish it directly.

continue to call the callback for now to support existing users.