move ordered member id querying to a decorator for SolrDocument
instead of isolating this logic `MemberPresenterFactory`, where i need an
Ability and a request context in order to run the query, extract for use with
`SolrDocument`.

since this isn't behavior for accessing `SolrDocument` data, per se, we can
avoid clutter and accidental querying by putting it in a decorator. to use, you
need to explictly `.decorate` the `SolrDocument` instance.

this will actually work for anything that's indexed with an `#id` and that has
`ActiveFedora`-style list proxy indexes. that means it *won't* work for
`Valkyrie` resources, at least as currently setup. maybe we don't need it for
them? work still to do on order there.