reload valkyrie resource after running after_create
If the resource is changed in the after_create, those changes are lost unless result is reloaded with those changes.  This uses the query_service to find and reload the resource.

Example where this is required, in after_create for a work, it is added to a collection.  The act of adding the work to the collection changes the resource.  The updated work resource needs to be returned by the ValkyrieCreateStrategy for the test and methods it calls to pick up the changes to the resource.