allow for extensions of valkyrie indexer's to_solr that depends on resource custom metadata
Fixes #4464

Allow shared spec testing of indexers with customizations to #to_solr that depend on properties defined in the resource class that the indexer indexes.

Update shared indexers spec to have a valid resource passed in to the shared spec.  The shared spec sets values it depends on in a before block.