add a test app: `.dassie`
it's useful to have a test application in the repository. this is something most
rails engines maintain. there's nothing stopping us from continuing to generate
test applications across rubies/rails versions, etc... but having an application
in the repo will help us to clearly manage upgrade processes, etc...

naming it "dassie" which is another name for a Hyrax.