alternate ids should be string-indexed-multivalue
Solr documents generated through valkyrie indexing cannot be saved because schema does not support _sm extension used with alternate_ids.

The extension should instead be _sim which makes it a searchable multi-valued string.  With this extension, it will not be part of the returned solrdoc when searching.  Assuming that is the desired behavior, then the _sim extension is the correct one to use.