dassie: support memcached cache store
add `dalli` to the application `Gemfile` for memcached support and configure
`:mem_cache_store` on the following conditions:

  - development caching is turned on;
    - this can be toggled in docker with `docker-compose exec app rails
      dev:cache`;
  - a `MEMECACHED_HOST` is provided;
    - if no host is given, but chaching is on, we fall back on `:memory_store`.

`MEMECACHED_HOST` is set via `.dassie/.env`, which is a newly added env
file. we make this file separate so we have a home for env configuration that
only relates to the `.dassie` application. variables that aren't present in the
engine can be slotted here to avoid confusion.