Introducing where clause for number_of_works()
Based on some investigation around #3478, I suspect that the
`Hyrax::WorkRelation#where` call that uses only a depositor field and
user key is finding more results than the query run on the catalog
search page.

By adding the where clause, I hope to align the query logic of the
`number_of_works` and the `link_to_field` on the [_vitals.html.erb][1] partial.

Related to #3478

[1]: https://github.com/samvera/hyrax/blob/b04c41d2772505d8a548c3fd48774a5642d3ac90/app/views/hyrax/users/_vitals.html.erb#L11-L12