Adding where clause to align counts on two pages
On the "vitals" page, there is a count that we derive from
`number_of_works`.  When you click on the `link_to_field` that new page
also as a count based on the provided facets.  Prior to this commit, I
believe the underlying queries were not using the same logic.

With this change, I hope to bring this in alignment.

Please note, I don't have tests to confirm this, nor a data state that
would help confirm it.  So, I'm pushing up a change that I believe
should work.

My plan is to write some helper tests at a later date, but for now am
drawing to the end of my work day.

Related to #3478