Use response code from Hyrax::API
The response_type used for 500 errors is `:internal_error` which doesn't map to a valid rails response type (it should be `:internal_server_error`).  Instead of changing Hyrax::API just use the code set in Hyrax::API.  There is probably a better fix for this but this avoids errors for now.