Recommended updates.
The above title is from the original commit.  What follows is the
description put in the PR message:

Add parent_presenter to file_set_presenter and use it to enforce access
restrictions from workflow mediated review. If public view not available
generate an unavailable page similar to a work's unavailable page.

Fixes #3775 ; refs #3881

Update to prevent direct navigation to file sets under works that are
public but restricted by workflow. When a user navigates to public file
sets in a workflow without and that user does not have edit privileges,
an unavailable screen is displayed.  From a work in the workflow that is
pending review, the file sets will not be
editable/deletable/downloadable.  Direct navigation to a file set for
users without edit privileges will result in an unavailable screen.

Guidance for testing, such as acceptance criteria or new user interface
behaviors:

* As a non-admin user, create a work with an attached file and place it
  in a mediate workflow (such as one step mediated workflow)
* From the work, you should not be able to edit/delete/download the file
* Navigate to the file set, it will show an unavailable file set
* Copy the url of file set
* Log out
* Directly navigate to the file set url, it will show unavailable