Compressing logic around workflow restriction
This commit attempts to normalize and consolidate the underlying logic
introduced in the Hyrax::FileSetsController.

As commented, there are application states in which we call the
guard_for_workflow_restriction_on! method:

- show action: when we start from a `presenter`
- edit/update/delete: when we start from a `curation_concern`

Both of those objects have a `parent` object but of different types.
The guard_for_workflow_restrion_on! now handles both of those parent
object types.

In doing so, the method begins to look like another method introduced in
a presenter (as noted).

To ensure that this works in the controller, I added the `#suppressed?`
method to the WorkShowPresenter.  Perhaps an oversight in delegation.