move indexing adapter registration to high-precedence
applications will probably want to do indexer setup in
initializers. registration of Hyrax's shipped adapters needs to be in place
before they do this. previously, this was the case only when the application
does setup in a letter after `'i'` (e.g. not `hyrax.rb`).

moving this up in precedence will make things more predictable for apps.