Synchronizing Spec::SpyListener and Publisher
I added documentation to explain how to add a named listener, instead of
an anonymous block listener.  I think this is useful, as our SpyListener
used a named listener, but that mechanism was not documented in the
Hyrax::Publisher.

In addition to the documentation, I've ensured that the SpyListener's
methods are synchronized with the Hyrax::Publisher registered events.
This synchronization also provides a bit of a test of the implementation
details of Dry::Events; Based on Dry's long-standing volatilitaty of
pre-1.0 gems, I think its a good plan to test a bit of the edges of
Dry::Events.