Ubiquity auto expiry jobs (#4502)
* Adds jobs for Expriy of Embargos

* Adds jobs for Expiry of Leases and cleans up specs

Fixes Rubocop issues

Fixes lease objects

Fixes lease objects

Fixes Typo

* Refactors expiry into a single job

Fixes Rubocop errors

Fixes more Rubocop errors

* move indexing adapter registration to high-precedence

applications will probably want to do indexer setup in
initializers. registration of Hyrax's shipped adapters needs to be in place
before they do this. previously, this was the case only when the application
does setup in a letter after `'i'` (e.g. not `hyrax.rb`).

moving this up in precedence will make things more predictable for apps.

* Synchronizing Spec::SpyListener and Publisher

I added documentation to explain how to add a named listener, instead of
an anonymous block listener.  I think this is useful, as our SpyListener
used a named listener, but that mechanism was not documented in the
Hyrax::Publisher.

In addition to the documentation, I've ensured that the SpyListener's
methods are synchronized with the Hyrax::Publisher registered events.
This synchronization also provides a bit of a test of the implementation
details of Dry::Events; Based on Dry's long-standing volatilitaty of
pre-1.0 gems, I think its a good plan to test a bit of the edges of
Dry::Events.

* Update app/jobs/embargo_expiry_job.rb

Co-authored-by: Tom Johnson <tomjohnson@ucsb.edu>

* Update app/jobs/lease_expiry_job.rb

Co-authored-by: Tom Johnson <tomjohnson@ucsb.edu>

* Bumping regen to fix rubocop-ast caching issue

Prior to this commit, the CI environment of Hyrax went wonky.  This, we
hope, is addressed in our Bixby implementation (see
https://github.com/samvera-labs/bixby/pull/47).  However, we need to
invalidate the cache.

* Adds jobs for Expriy of Embargos

* Adds jobs for Expiry of Leases and cleans up specs

Fixes Rubocop issues

Fixes lease objects

Fixes lease objects

Fixes Typo

* Refactors expiry into a single job

Fixes Rubocop errors

Fixes more Rubocop errors

Co-authored-by: tamsin johnson <tomjohnson@ucsb.edu>
Co-authored-by: Jeremy Friesen <jeremy.n.friesen@gmail.com>
Co-authored-by: E. Lynette Rayle <elrayle@users.noreply.github.com>