index valkyrie work attributes consistent with AF
Most attributes for Hyrax::Work are defined using `Hyrax::Schema` includes.  A few are defined with `attribute` statements.  Some of these attributes were indexed by ActiveFedora.  Adding in indexing for these to be consistent with ActiveFedora’s indexing.

Indexed in AF, but not in ValkyrieWorkIndexer…
* generic_type_sim
* file_set_ids_ssim

Defined as attribute in Hyrax::Work, but not indexed
* on_behalf_of
* proxy_depositor
* state

Looking for confirmation on how to handle these.

Issue #4487 - Solr docs generated by ActiveFedora and Hyrax::ValkyrieWorkIndexer are not compatible