Restoring HTML form field name.
In 241cbed1df4ad416b8d44763ae321e642443ed88 we introduced changes to
address accessibility issues reported in #3966.  However, the introduced
change broke the HTML form field names and the assumed parameters for
the [batch processing][1].

This commit reverts that change, but it does not address the
accessibility concern raised in #3966.  That is for another commit, as I
don't know if that other commit is the correct path forward.

Closes #4516
Closes #4515

[1]:https://github.com/samvera/hyrax/blob/3c7258c95c1fe6f0fe1537078f3f5f458f211989/app/controllers/concerns/hyrax/collections/accepts_batches.rb#L20-L28