Adding fieldset to address like named form inputs
In consultation with @dananji and following the [guidance of W3.org][1],
this commit wraps the forms (or controls) that render the
[add_buttom partial][2].  I don't believe that I have access to the
accessibility testing software, so this is a bit of a stab in the dark.

Note, in some cases I render the fieldset as a direct child of the form,
in other cases, I render the fieldset as a container for the table used
to present the elements that render the add button partial.

This relates to work done in #4160 to address #3966.

[1]:https://www.w3.org/WAI/tutorials/forms/grouping/
[2]:https://github.com/samvera/hyrax/blob/d5aa2f9ca802fe670687bb4f7c14ad242c1bacf6/app/views/hyrax/batch_select/_add_button.html.erb#L2