Adding feature test for Collection#destroy
A little bit about this spec.  It's informed by #3762 which is why it
looks a bit odd.  When I wrote this spec it worked.  I suspect that some
prior change between the reported date of June 28, 2019 and
today (October 8, 2020) fixed this bug.

See [Hyrax::BatchEditsController#destroy_collection][1] for the related
code that informed the spec's implementation details.

I'm not prepared nor willing to bisect when this bug was fixed.  I also
understand that we need not merge this commit as the code-base already
works.  I do, however, think about some of the assumptions that we've
found to be broken; [Looking at you Embargo and Lease cascade behavior][2].

Closes #3872

[1]:https://github.com/samvera/hyrax/blob/0e6e7c4ff72fe508d2af0bb2eb22a1a532cf54b3/app/controllers/hyrax/batch_edits_controller.rb#L40-L47
[2]:https://github.com/samvera/hyrax/issue/4534