Ensuring embargo and lease are persisted on update
Prior to this commit, when a user changed the Embargo or Lease in the
UI (and by extension via the ActorStack), Hyrax would not persist those
changes.

With this change, we ensure that when the visibility changes, hyrax
updates the embargo and lease.

In consultation with @no_reply, we believe that this should be handled
further upstream (e.g. within `Hydra::AccessControls`).  We also believe
that there was an upstream change that broke the previous correct
behavior.

We are attempting to track that down.

Closes #4534, #4517