Ensuring AdminSet properly indexed during seeds
I have found that when I come back to a development environment, that I
may have an AdminSet in Fedora, but it is not indexed in Solr.  This
remediates that situation by ensuring we have an indexed AdminSet.

We [merged this change merged][1] into `rake
hyrax:default_admin_set:create` but it was not part of the db:seeds.

[1]:https://github.com/samvera/hyrax/commit/940e1003e35b55a1aab69cdc2a231269fe7f29bf