Demoting all README headers
Prior to this commit, the rendered Markdown creates many H1-tags; That
is against best practices for HTML. There should be one H1 tag per
document, or at a minimum per `<section>` tag.

This change introduces one H1 and the rest of the headers are
sub-headers. This should help improve assistive technology navigate the page.

[skip ci]