Refining README and CONTAINERS for current state
Let's be honest, we have many paths to test, develop, and deploy these
applications.  And given our limited resources, we need to commit to a
repeatable/reusable direction.

However, it may not be the best direction for everyone.  Ergo, you get
the "we're trying" and here's some places to look.