wings: treat FileMetadata like any other resource
there are a few special things about `FileMetadata`. it needs to map to an
`AF::File::MetadataNode` (these are a bit complicated themselves), which can't
be created independently of a `AF::File `, some of their metadata (e.g. `type`)
has special handling in `Hydra::PCDM` and lower.

but for the most part, these are the same as any other Resource. instead of
implementing a special conversion service for them, just work their special
handling into the existing infrastrucure. for the most part, this doesn't even
require new abstractions. hopefully this is cleaner, and will encourage this
code to be better understood and maintained.

`FileMetadataBuilder` is next.