wings: deprecate FileMetadataBuilder
you can now use `storage_adapter.upload` to do all the work of
`FileMetadataBuilder`. Fedora creates the metadata node automatically, so it
just needs to be queried.

We may still want a Hyrax-layer abstraction for handling the assumption that
there will always be a `FileMetadata` model for every stored (pcdm) File. This
would need special handling for non-Wings adapters, but not for Wings.