extract and minimize valkyrie fileset attachment behavior
preparing to do fileset attachment outside the actor stack context. this is the
simplest version of valkyrie driven fileset attachment that i could come up with
that passes the current tests.

i fear that passing the current tests may not be enough. but i think the
preferred strategy at this point is to get an end-to-end implementation from the
`WorksControllerBehavior` tests so we can avoid duplicating the byzantine
processes involved in FileSetActor, FileActor, AttachFileToWorkJob, etc...