implement transaction interface for file set attachment
when creating a work, upload file sets if step args are passed.

adds a new transaction step that calls an (injectable) handler to create
FileSets and enqueue file uploads.

for now, this attaches file sets inline. this might not be sustainable (the
actor stack enqueues jobs), but it should let us test things
end-to-end. replacing with a version that enqueues a job `AddFileSetsAsync` is
fairly easy with the dry-container interface, so it seems best to defer.