remove blanket stubbing from `generic_works_controller_spec#update`
these tests over-isolated the Controller. at some point along the way (around
2017; 94c8663058df2b16c646c5de112ee5c5c1932a4b) they stopped stubbing realistic
GenericWork behavior around `visibility_changed?`. to fix this, and reduce the
risk of further drift, back out the stubs and test some realistic interactions.

we keep a few isolated stubs, not for any principled reason, but just to reduce
the footprint of this change. probably stubbing with a smaller scope will be
easier to maintain?