document `WorkUploadsHandler` idempotency/retryability
`WorkUploadsHandler`, like the implementation it extracted, is *technically*
idempotent. it achieves this by simply ignoring uploads with partially complete
work, so it's not actually retry-able. bad news for `AttachFileToWorkJob`, we
should fix it later.

the `raise` added isn't ever actually reached, but `WorkUploadsHandler`'s
documented API says suggests it will return falsey on `#apply` if file
attachments fail. in practice we raise immediately, but the caller should handle
the false case.