set depositor on create with Valkyrie models
we infer the depositor from application context.

the old solution was to push the whole context into the form object. this led to
a mass of interdependencies between form, controller, and model. hoping to avoid
this with Valkyrie/Reform, but we still have the original problem: how to get
controller context into the process.

the two options that were considered are:

  - Reform's `#prepopulate(options)`
  - Transaction's `#with_step_args`

the main reasons for going with the latter here are:

     -  we're already using the feature to pass controller data into the create
     process
     - the transaction step already exists and supports the needed argument
     - we call `#prepopulate` from `FormFactory`, and we're trying also to
     reduce collaborators there.

it might be that we find `Hyrax::FormFactory` really *needs* either the Ability
or the Controller for context (i say in only ~wants~ it in this instance). if we
get there, we might want reconsider this approach.