introduce a listener to handle proxy deposits; deprecate TransferRequestActor
in the past, we've enqueued `ContentDepositorChangeEventJob` from an
actor (`TransferRequestActor`). this is pretty limited, since it means any proxy
deposit MUST go through the actor stack, and that actor MUST be triggered.

since this behavior is a reaction do an object's deposit, it's a good candidate
for a listener. we already promise to publish events for
`'object.deposited'`. adding a listener as the handler means we are able to
respond to any deposit. since listeners are "dead ends", with no downstream
contract, creating a custom listener with different behavior (e.g. handle this
transfer synchronously, trigger applicaiton-specific events or notifications,
etc...) should be easier than creating a custom actor.

we choose to enqueue a job if the depositor has deposited "on behalf of" a
non-existant user. this is a choice, to be sure, but it seems better than either
panicking, or just logging. the job will (presumably) fail (unless the missing
user was a fleeting error), and in most deployments get queued for retry. this
gives local maintaniers a chance to create the user, if that's what should have
happened, or remove the job from the queue if they prefer a no-op.

**breaking change**: this DEPRECATES `TransferRequestActor` and REMOVES it from
  the default actor stack. existing default actor stack behavior isn't changed,
  but downstream customizations to `TransferRequestActor` require
  application-side action to retain. applications can either port their
  customizations to the listener, or defer by unsubscribing the listener and
  re-inserting the deprecated actor.