support valkyrie in WorkflowFactory
generalizes Sipity workflow initialization to work with Valkyrie::Resource
models.

this doesn't pretend to solve workflows for Valkyrie in total, but does ensure
existing expectations for the `WorkflowFactory` initialization process are
portable.

the main obvious gap is that `Hyrax::Workflow::WorkflowActionService` has
a (undesirable?) dependency on `ActiveFedora::Base#update_index`. what is it
counting on being indexed? i'm not yet sure; but it's unlikely we've addressed
it in the valkyrie indexers. this will need some attention as we continue to
work on Valkyrie workflows.