use Workflow::Listener to initialize workflow
we want to initialize workflow on deposits in general, not just through the
actor stack.

*breaking change*: this DEPRECATES `InitializeWorkflowActor` and REMOVES it from
the default actor stack. while the replacement functionality should be an
improvement over the older implementation, and is designed to respect certain
customizations (custom workflow factories set via the existing class
configuration are retained), application-side action is required if a custom
actor is used.